refactor(log): split LogPanel.tsx into log-panel-model, useFollowTail, LogEntryRow, shared observe-element-offset
C-16: LogPanel (754L) decomposed into:
- lib/observe-element-offset.ts: shared virtualizer util (leak fix)
- app/log-panel-model.ts: filter/severity/navigation pure functions
- app/useFollowTail.ts: scroll follow-tail state machine hook
- app/LogEntryRow.tsx: single virtualized row component

LogPanel.tsx retains composition only. Behavior-preserving, typecheck green.
